ml/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
alembic==1.13.1
pydantic==2.7.4
pydantic-settings==2.3.3
pymysql==1.1.1
aiomysql==0.2.0
SQLAlchemy==2.0.31
pyjwt==2.8.0
cryptography==42.0.8
bcrypt==4.2.0
